[UI] Phase 4: move MesheryFilters to ui/components/filters (#18749)
Rename ui/components/MesheryFilters/ to ui/components/filters/ to align
with the flatter, domain-oriented structure defined in the Phase 4
restructure plan (move order item 10). Pure structural rename via
git mv, no behavior changes.

- git mv ui/components/MesheryFilters ui/components/filters
- Update lone importer in pages/configuration/filters.tsx
- Refresh eslint.config.js legacy allowlist paths (color literals,
  max-line, and inline-style offender lists) so allowlist entries
  continue to match the renamed files

Fixes #18749
Refs #18659, #18654
