+++ ui/eslint.config.js
@@ -98,8 +98,6 @@ const legacyLiteralColorOffenders = [
   'components/MeshAdapterConfigComponent.tsx',
   'components/MesheryAdapterPlayComponent.tsx',
   'components/MesheryChart.tsx',
-  'components/MesheryFilters/Filters.tsx',
-  'components/MesheryFilters/FiltersGrid.tsx',
   'components/MesheryMeshInterface/PatternService/RJSFCustomComponents/ArrayFieldTemlate.tsx',
   'components/MesheryMeshInterface/PatternService/RJSFCustomComponents/CustomBaseInput.tsx',
   'components/MesheryMeshInterface/PatternService/RJSFCustomComponents/ObjectFieldTemplate.tsx',
@@ -137,6 +135,8 @@ const legacyLiteralColorOffenders = [
   'components/connections/meshSync/Stepper/StepperContentWrapper.tsx',
   'components/connections/meshSync/Stepper/index.tsx',
   'components/connections/styles.tsx',
+  'components/filters/Filters.tsx',
+  'components/filters/FiltersGrid.tsx',
   'components/load-test-timer-dialog.tsx',
   'components/telemetry/grafana/GrafanaCustomGaugeChart.tsx',
   'components/telemetry/grafana/GrafanaDateRangePicker.tsx',
@@ -151,10 +151,10 @@ const legacyMaxLineOffenders = [
   'components/Dashboard/resources/configuration/config.tsx',
   'components/Dashboard/resources/workloads/config.tsx',
   'components/MesheryAdapterPlayComponent.tsx',
-  'components/MesheryFilters/Filters.tsx',
   'components/MesheryPatterns/MesheryPatterns.tsx',
   'components/Performance/index.tsx',
   'components/connections/ConnectionTable.tsx',
+  'components/filters/Filters.tsx',
 ];
 
 // Files currently in the 600–1000 line "soft" range. They exceed the 600-line
@@ -240,10 +240,6 @@ const legacyInlineStyleOffenders = [
   'components/MesheryAdapterPlayComponent.tsx',
   'components/MesheryChart.tsx',
   'components/MesheryCredentialComponent.tsx',
-  'components/MesheryFilters/CatalogFilter.tsx',
-  'components/MesheryFilters/Filters.tsx',
-  'components/MesheryFilters/FiltersCard.tsx',
-  'components/MesheryFilters/FiltersGrid.tsx',
   'components/MesheryMeshInterface/PatternService/RJSFCustomComponents/Accordion.tsx',
   'components/MesheryMeshInterface/PatternService/RJSFCustomComponents/ArrayFieldTemlate.tsx',
   'components/MesheryMeshInterface/PatternService/RJSFCustomComponents/CustomBaseInput.tsx',
@@ -337,6 +333,10 @@ const legacyInlineStyleOffenders = [
   'components/connections/meshSync/index.tsx',
   'components/connections/metadata.tsx',
   'components/extensions/adapters/adapters.tsx',
+  'components/filters/CatalogFilter.tsx',
+  'components/filters/Filters.tsx',
+  'components/filters/FiltersCard.tsx',
+  'components/filters/FiltersGrid.tsx',
   'components/layout.tsx',
   'components/multi-select-wrapper.tsx',
   'components/navigatorComponents.tsx',
